Decouple data-renewal from the concrete EntryKind via a RenewMeta bound
