+++ runtimes/bulletin-westend/Cargo.toml
@@ -45,6 +45,7 @@ pallet-bulletin-data-renewal = { workspace = true }
 pallet-bulletin-transaction-storage = { workspace = true }
 pallet-bulletin-transaction-storage-runtime-api = { workspace = true }
 bulletin-pallets-common = { workspace = true }
+bulletin-transaction-storage-primitives = { workspace = true }
 sp-api = { workspace = true }
 sp-block-builder = { workspace = true }
 sp-consensus-aura = { workspace = true }
@@ -91,7 +92,6 @@ testnet-parachains-constants = { features = ["westend"], workspace = true }
 [dev-dependencies]
 parachains-runtimes-test-utils = { workspace = true, default-features = true }
 sp-io = { workspace = true }
-bulletin-transaction-storage-primitives = { workspace = true }
 
 [build-dependencies]
 substrate-wasm-builder = { optional = true, workspace = true, default-features = true }
